Improve test support
* Add some documentation
* All values are checked depending on their types
* Do not compute Python output if check_output is False
* Do not hide misspelling in spec parsing for tests
* Accept '#runas' and '# runas' (like for export) (pep8)
* PEP8

+++ pythran/pythonic/numpy/around.hpp
@@ -22,10 +22,10 @@ namespace pythonic {
         template<class E>
             auto around(E const& a, long decimals=0)
             -> typename std::enable_if<std::is_integral<typename types::numpy_expr_to_ndarray<E>::type::dtype>::value,
-               decltype(a / nt2::pow(typename types::numpy_expr_to_ndarray<E>::type::dtype(10),std::max(0L,-decimals)) * nt2::pow(typename types::numpy_expr_to_ndarray<E>::type::dtype(10), std::max(0L,-decimals)))
+               decltype(a / (long)nt2::pow(typename types::numpy_expr_to_ndarray<E>::type::dtype(10),std::max(0L,-decimals)) * (long)nt2::pow(typename types::numpy_expr_to_ndarray<E>::type::dtype(10), std::max(0L,-decimals)))
                  >::type
             {
-                return a / nt2::pow(typename types::numpy_expr_to_ndarray<E>::type::dtype(10),std::max(0L,-decimals)) * nt2::pow(typename types::numpy_expr_to_ndarray<E>::type::dtype(10), std::max(0L,-decimals));
+                return a / (long)nt2::pow(typename types::numpy_expr_to_ndarray<E>::type::dtype(10),std::max(0L,-decimals)) * (long)nt2::pow(typename types::numpy_expr_to_ndarray<E>::type::dtype(10), std::max(0L,-decimals));
             }
 
         template<class T>
